merge develop into main

+++ src/stores/controllers/stores.controller.ts
@@ -6,13 +6,20 @@ import {
   Param,
   Delete,
   Query,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { StoresService } from '../services/stores.service';
 import { StoreDtoRequest } from '../dtos/requests/store-request.dto';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { StoreDtoResponse } from '../dtos/responses/store-response.dto';
+import { AuthGuard } from 'src/guards/auth.guard';
+import { CustomerTypeEnum } from 'src/common/enums/customer-type.enum';
+import { IsLogist } from 'src/decorators/logist.decorator';
+import { LogistDtoResponse } from 'src/logists/dto/responses/logist-response.dto';
 
 @ApiTags('Stores')
+@UseGuards(AuthGuard)
 @Controller('api/v1/stores')
 export class StoresController {
   constructor(private readonly storesService: StoresService) {}
@@ -24,8 +31,15 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 400, description: 'Invalid request' })
-  create(@Body() dto: StoreDtoRequest): Promise<StoreDtoResponse> {
-    return this.storesService.create(dto);
+  create(
+    @IsLogist() logist: LogistDtoResponse,
+    @Body() dto: StoreDtoRequest,
+  ): Promise<StoreDtoResponse> {
+    if (logist.customerType === CustomerTypeEnum.ADMIN) {
+      return this.storesService.create(logist.id, dto);
+    } else {
+      throw new ForbiddenException('Access denied.');
+    }
   }
 
   @Get('by-user')
@@ -35,8 +49,13 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  findAllByUser(): Promise<StoreDtoResponse[]> {
-    return this.storesService.findAllByUser();
+  findAllByUser(
+    @IsLogist() logist: LogistDtoResponse,
+  ): Promise<StoreDtoResponse[]> {
+    if (logist.customerType === CustomerTypeEnum.ADMIN) {
+      return this.storesService.findAllByUser(logist.id);
+    }
+    throw new ForbiddenException('Access denied.');
   }
 
   @Get('by-store')
@@ -46,14 +65,26 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  searchById(@Query('storeId') storeId: string): Promise<StoreDtoResponse> {
-    return this.storesService.searchById(+storeId);
+  searchById(
+    @IsLogist() logist: LogistDtoResponse,
+    @Query('storeId') storeId: string,
+  ): Promise<StoreDtoResponse> {
+    if (logist.customerType === CustomerTypeEnum.ADMIN) {
+      return this.storesService.searchById(logist.id, +storeId);
+    }
+    throw new ForbiddenException('Access denied.');
   }
 
   @Delete(':storeId')
   @ApiResponse({ status: 204, description: 'Remove store' })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  remove(@Param('storeId') storeId: string) {
-    return this.storesService.remove(+storeId);
+  remove(
+    @IsLogist() logist: LogistDtoResponse,
+    @Param('storeId') storeId: string,
+  ) {
+    if (logist.customerType === CustomerTypeEnum.ADMIN) {
+      return this.storesService.remove(+storeId);
+    }
+    throw new ForbiddenException('Access denied.');
   }
 }
